# openapi/coinbase-admin-api.yaml
openapi: 3.1.0
info:
  title: Coinbase Admin API
  description: Enterprise API for the Coinbase trading engine. Read-only operator
    routes are active; mutating HTTP routes use shared command services but remain
    live-disabled by the approval gate.
  version: 0.1.0
paths:
...
  /api/v1/orders:
    get:
      tags:
      - orders
      summary: Read local orders keyed by client_order_id
      description: Read local order_parent evidence without contacting Coinbase.
      operationId: list_orders_api_v1_orders_get
      parameters:
      - name: product_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Id
      - name: order_status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Order Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderListResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - orders
      summary: Create a manual order through the shared command service
      description: 'Route adapter for manual placement.


        The route is authenticated, idempotent, audited, and still live-disabled

        until enterprise approval/cap gates are completed.'
      operationId: create_manual_order_api_v1_orders_post
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          title: Idempotency-Key
      - name: X-Correlation-Id
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          title: X-Correlation-Id
      - name: X-Operator-Intent
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          title: X-Operator-Intent
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualOrderRequest'
      responses:
        '200':
          description: Command accepted or replayed after all backend gates pass.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '400':
          description: Command rejected before live exchange execution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '409':
          description: Idempotency key conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '501':
          description: Live HTTP execution is not implemented for this command.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{client_order_id}:
    get:
      tags:
      - orders
      summary: Read one local order by client_order_id
      description: Read one local order row by client_order_id.
      operationId: get_order_by_client_order_id_api_v1_orders__client_order_id__get
      parameters:
      - name: client_order_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          title: Client Order Id
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderDetailResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{client_order_id}/cancel:
    post:
      tags:
      - orders
      summary: Cancel an order by client_order_id through the shared command service
      description: Route adapter for cancel-by-client-order-id.
      operationId: cancel_order_by_client_order_id_api_v1_orders__client_order_id__cancel_post
      parameters:
      - name: client_order_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          title: Client Order Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          title: Idempotency-Key
      - name: X-Correlation-Id
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          title: X-Correlation-Id
      - name: X-Operator-Intent
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          title: X-Operator-Intent
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelOrderRequest'
      responses:
        '200':
          description: Command accepted or replayed after all backend gates pass.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '400':
          description: Command rejected before live exchange execution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '409':
          description: Idempotency key conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '501':
          description: Live HTTP execution is not implemented for this command.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spot/campaign/executions:
    post:
      tags:
      - orders
      summary: Execute a spot campaign through the shared command service
      description: 'Route adapter for future campaign execution.


        The route has the command envelope, idempotency, audit, RBAC, and fail-closed

        live gate, but it does not submit Coinbase orders.'
      operationId: execute_spot_campaign_api_v1_spot_campaign_executions_post
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          title: Idempotency-Key
      - name: X-Correlation-Id
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          title: X-Correlation-Id
      - name: X-Operator-Intent
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          title: X-Operator-Intent
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignExecutionRequest'
      responses:
        '200':
          description: Command accepted or replayed after all backend gates pass.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '400':
          description: Command rejected before live exchange execution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '409':
          description: Idempotency key conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '501':
          description: Live HTTP execution is not implemented for this command.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiCommandResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AdminApiCommandResponse:
      properties:
        status:
          $ref: '#/components/schemas/AdminApiCommandStatus'
        action_class:
          $ref: '#/components/schemas/AdminApiActionClass'
        required_permission:
          $ref: '#/components/schemas/AdminApiPermission'
        service_method:
          type: string
          title: Service Method
        message:
          type: string
          title: Message
        client_order_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Order Id
        coinbase_order_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Coinbase Order Id
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
        audit_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Audit Id
        live_exchange_submitted:
          type: boolean
          title: Live Exchange Submitted
          default: false
        submission_event_recorded:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Submission Event Recorded
        audit_command:
          anyOf:
          - type: string
          - type: 'null'
          title: Audit Command
        guard:
          anyOf:
          - type: object
          - type: 'null'
          title: Guard
        data:
          anyOf:
          - {}
          - type: 'null'
          title: Data
        failure_stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Failure Stage
      additionalProperties: false
      type: object
      required:
      - status
      - action_class
      - required_permission
      - service_method
      - message
      title: AdminApiCommandResponse
      description: Typed response returned by Admin API command adapters.
...
    AdminApiErrorCode:
      type: string
      enum:
      - auth_required
      - permission_denied
      - validation_error
      - idempotency_conflict
      - guard_blocked
      - not_implemented
      - backend_unavailable
      - request_error
      title: AdminApiErrorCode
      description: Structured Admin API error codes exposed to the frontend.
    AdminApiErrorResponse:
      properties:
        code:
          $ref: '#/components/schemas/AdminApiErrorCode'
        message:
          type: string
          title: Message
        severity:
          $ref: '#/components/schemas/AdminApiErrorSeverity'
        guard_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Guard Name
        field_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Path
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
        audit_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Audit Id
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: false
      type: object
      required:
      - code
      - message
      - severity
      title: AdminApiErrorResponse
      description: Structured error body shared by Admin API routes.
...
    AdminOrderListResponse:
      properties:
        type:
          type: string
          title: Type
          default: admin_order_list
        filters:
          type: object
          title: Filters
        count:
          type: integer
          title: Count
        items:
          items:
            $ref: '#/components/schemas/AdminOrderReadItem'
          type: array
          title: Items
        read_only:
          type: boolean
          title: Read Only
          default: true
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: false
      type: object
      required:
      - count
      title: AdminOrderListResponse
      description: Read-only order list/filter response.
...
    CancelOrderRequest:
      properties:
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      additionalProperties: false
      type: object
      title: CancelOrderRequest
      description: Cancel request body keyed by path ``client_order_id``.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ManualOrderRequest:
      properties:
        product_id:
          type: string
          minLength: 1
          title: Product Id
          examples:
          - BTC-USDC
        side:
          $ref: '#/components/schemas/OrderSide'
        order_type:
          $ref: '#/components/schemas/OrderType'
        base_size:
          anyOf:
          - type: string
            pattern: ^-?(0|[1-9]\d*)(\.\d+)?$
            description: Decimal value serialized as a string; floats are not part
              of the API contract.
            examples:
            - '1.00'
          - type: 'null'
          title: Base Size
        quote_size:
          anyOf:
          - type: string
            pattern: ^-?(0|[1-9]\d*)(\.\d+)?$
            description: Decimal value serialized as a string; floats are not part
              of the API contract.
            examples:
            - '1.00'
          - type: 'null'
          title: Quote Size
        limit_price:
          anyOf:
          - type: string
            pattern: ^-?(0|[1-9]\d*)(\.\d+)?$
            description: Decimal value serialized as a string; floats are not part
              of the API contract.
            examples:
            - '1.00'
          - type: 'null'
          title: Limit Price
        post_only:
          type: boolean
          title: Post Only
          default: false
        time_in_force:
          anyOf:
          - $ref: '#/components/schemas/TimeInForce'
          - type: 'null'
        manual_live_acknowledgement:
          type: boolean
          title: Manual Live Acknowledgement
          default: false
      additionalProperties: false
      type: object
      required:
      - product_id
      - side
      - order_type
      title: ManualOrderRequest
      description: Manual order request shape for future enterprise placement.
...